Admin Login
Admin login now works, need to add the emails and passwords of JM and Keith to the Firebase User List to login the website.

Using Firebase Authentication for login feature.

Once Admin is logged in, navbar displays an admin panel accordion that displays list of links to click.

Accordion is also where the logout button is located.