Created Edit Announcement, Created Delete Announcement
Created Edit Announcement page to fetch announcement from DB and display the data.

Created Delete Announcement Function to delete data from the website and the DB. imagePath for images also added when being added/edited so when deleting, it will delete in the back end ass well so no orphan image files.

Next to work on is having the feature to click announcement cards and have a pop up of the announcement for better visibility. Card pop up should be kind of like an overlay of the page, off canvas kinda.